+++ packages/visual-editor/package.json
@@ -290,7 +290,6 @@
     "vitest": "^2.1.3"
   },
   "dependencies": {
-    "@bjorn3/browser_wasi_shim": "^0.3.0",
     "@breadboard-ai/board-server-management": "1.18.0",
     "@breadboard-ai/board-server-utils": "0.1.4",
     "@breadboard-ai/build": "0.10.4",
